simple genetic algorithms implementation
A basic implementation of genetic algorithms (GA) in Python3. GA is a
metaheuristic inspired by the process of evolution.
Shows the changes in genotypes (expressed by a sequence Boolean values
in an array) over multiple generations,
matching the best-fitting genotype determined by an environment.

We will need to model the following GA principals
- intializing a population with its genetic representation
- creating a fitness function
- creating a selection function
- creating a goal/termination point